refactor(adapter-openclaw): extract the cron store into cron-store.ts
Third cut of the runtime.ts god-file split. Moves the entire runtime.cron
persistence + parsing layer — 23 helpers + the 3 store types
(OpenClawCronStore/Job/RunEntry) + OPENCLAW_CRON_TIMEOUT_MS — into a new
cron-store.ts. Covers jobs.json read/write, runs/<id>.jsonl reads, the
`cron add|edit` CLI argv builders, and the loose-JSON normalizers that turn
OpenClaw CLI output into typed CronJob / CronRun shapes.

The class's cron methods keep the exec calls and import these helpers; the
module touches no adapter instance state, so it's a clean store/parse unit.
cron-store imports its leaf utils from runtime-utils (no cycle). runtime.ts
imports back the 16 helpers + the one type + the timeout const its cron
methods (and the agent-config cron sweep) still reference; the parse-only
internals stay module-private to cron-store.

Pure relocation — bodies byte-for-byte unchanged. runtime.ts 2,989 -> 2,600.

Verified: typecheck, eslint (changed files), adapter suite 115-0 incl.
runtime-cron.test.ts (--isolate), full suite 5124-0, binary build (stamps
reverted), and DOCKERIZED-RIG E2E (real OpenClaw, isolated mode): a native
cron created via the real CLI was listed through Bakin (runtime.cron.list ->
extractCronStoreJobs/normalizeOpenClawCronStoreJob/cronStoreJobToRuntime
parsed all 7 jobs; probe normalized to {type:cron, value:'*/5 * * * *'},
command from payload, enabled, isBakinJob:false), then removed via the CLI
and the re-list confirmed re-parse (7 -> 6). The moved code behaves
identically on the live stack.
